Line graph with dynamic data.

+++ app/src/main/java/com/valllent/websocket/data/Coin.kt
@@ -2,5 +2,5 @@ package com.valllent.websocket.data
 
 data class Coin(
     val name: String,
-    val price: Double
+    val price: Float,
 )
